Fix uploading large files from stream
Uploading bigger files uses an upload session to upload the file in smaller chunks. This session is initialized using the DropboxFile path.
When we create a new DropboxFile from a stream (using the "createyByStream" method), the path wasn't correctly set. The filename was set instead, so when we tried to upload a bigger file from a stream, the upload session tried to retrieve the file using the file name, resulting in an error.
This commit fixes this by setting the stream URI as the DropboxFile path.

Note: This'll only work for a stream that you get from tmpfile. For a regular file pinter, it might only give you the basename.